Support for HbbTV browsers (EU and AU Standard for Connected TVs)
Every year since 3 years, a dozen of TV model or so are released, each with its own capabilities. By chance the HbbTV standard has been adopted by European TV networks and manufacturers, and this standard also defines the User-Agent.
I tested the 2 regexes I added on a dozen of model and is using it with TestSwarm