Issue 297: Move Firefox patterns between IE11 and other IE
Sometimes IE11's UserAgent string contains Firefox/12 in it.  The
old pattern order results in this being identified as Firefox/12.

Also, old Firefox versions would include MSIE 6 in it.

In order to correctly identify IE11 and older Firefox, we need to
move the Firefox check after the IE11 check, but before all the
older IE checks.