Detect Chrome iOS when the Request Desktop Site feature is enabled
When the Request Desktop Site feature is enabled, Chrome on iOS only includes the major revision in the user agent. This change adds support for that and adds two test cases (examples of Chrome on iOS with and without the Request Desktop Site feature). See https://developer.chrome.com/docs/multidevice/user-agent/#chrome-for-ios